requirements.txt
jsonrpclib-pelix
newspaper3k
numpy
pika
pymongo
python-dateutil
redis
requests
scipy
sklearn
pandas
tensorflow
watchdog
